% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salso.R
\name{computeBinderDistance}
\alias{computeBinderDistance}
\title{Binder distance of two partitions}
\usage{
computeBinderDistance(testLabels, refLabels, Const_Binder)
}
\arguments{
\item{testLabels}{The vector of integer labels of the partitioning}

\item{refLabels}{The vector of partition labels to use to construct the adjacency matrix}

\item{Const_Binder}{Relative penalty in the Binder loss function 
for false-positives vis-a-vis false-negatives. 
Must be a real number in the interval [0, 1].}
}
\value{
The value of the Binder loss function of \code{testLabels} 
with respect to the adjacency matrix of \code{refLabels}.
}
\description{
Compute the binder loss function of a partitioning
with respect to the adjacency matrix of another partitioning.
}
