% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salso.R
\name{computeBinderLoss}
\alias{computeBinderLoss}
\title{Compute the Binder loss function}
\usage{
computeBinderLoss(eam, labels, Const_Binder = 0.5)
}
\arguments{
\item{eam}{Expected Adjacency Matrix, i.e., 
the matrix whose entries \eqn{E_{ij}} is the posterior probability 
that items \eqn{i} and \eqn{j} are together. 
If the partitioning is already known, this is just the adjacency matrix.}

\item{labels}{vector of partition labels. Must be integers.}

\item{Const_Binder}{Relative penalty in the Binder loss function 
for false-positives vis-a-vis false-negatives. 
Must be a real number in the interval [0, 1].}
}
\value{
The value of the Binder loss function of the given partition labels
with respect to the given pairwise allocation matrix.
}
\description{
Compute the Binder loss function of a partitioning
with respect to an expected adjacency matrix.
}
