% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salso.R
\name{salso}
\alias{salso}
\title{Sequentially Allocated Latent Structure Optimisation}
\usage{
salso(p, maxClust, Const_Binder = 0.5, nPerm = 10000L, nScans = 3L)
}
\arguments{
\item{p}{co-clustering probability matrix, also called the expected adjacency matrix}

\item{maxClust}{maximum number of clusters}

\item{Const_Binder}{relative penalty in the Binder loss function for false-positives vs false-negatives}

\item{nPerm}{number of random permutations to start optimising from}

\item{nScans}{number of scans for each permutation}
}
\value{
list containing a vector of cluster labels and the associated binder loss function
}
\description{
Heuristic clustering to minimise the expected Binder loss function with respect to a given co-clustering probability matrix.
}
